Fixed tests and minor changes after standalone test
