+++ shared/protocol/src/main/java/io/pravega/shared/protocol/netty/AdminRequestProcessor.java
@@ -1,0 +1,23 @@
+/**
+ * Copyright Pravega Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.pravega.shared.protocol.netty;
+
+/**
+ * A class that handles each type of Admin-specific Request.
+ */
+public interface AdminRequestProcessor extends RequestProcessor {
+    // Placeholder for new admin-specific requests to be added in the future.
+}
